edited: renamed the identity-verification directory and added the collection service
edited: renamed the identity-verification directory and added the collection service
