+++ src/services/index.ts
@@ -3,7 +3,7 @@ export * from './chargebacks/chargeback';
 export * from './conversions/conversion';
 export * from './quotes/quote';
 export * from './wallets/wallet';
-export * from './identity-verification/verify-bank';
+export * from './verification/verify';
 export * from './beneficiaries/beneficiary';
 export * from './payouts/payout';
 export * from './subaccounts/subaccount';
